HC-58: Add API Wrapper to Change Params Array Structure
A bug injected on CiviCRM v4.7.19 caused chained API calls using '$value.' to
chain values to previous call to be interpreted as operators with syntax:
array('OP' => $val). This only happened if the first parameter for the chained
API call uses '$value.'.

Fixed by implementing civicrm_apiWrappers hook to preprocess parameters array
to make sure the first parameter of a chained API call is is not a chained
value.