HC-58: Handle Case Where All Parameters are Chained Values
As what the implemented fix does is move chained values to the last position
of the array, if all the parameters of the call are chained values, the
problem would still present itself.

Added check to verify if all values are chained and if so, added sequential
flag with 0 value as parameter.