HC-61: Fix Configuration Selection Combo on Basket Add Form
Selection combo was showing all configuration options, no matter what resource
was selected. This was happening because th configuration set's ID was not
being used to filter the result. On the other hand, the option's label was
being used instead of the unit's label where units hould appear.

Fixed the API call so config set's ID was properly chained to options. Changed
the template for config options so unit label is used instead of config option
label.