HC-61: Fix Loading of Resource Configuration Options
Form to add unlimited resources got broken when the template used to generate
the options was changed to include units, as the API call to obtain options
did not use the sequential flag. This is important, because the unit label is
being accessed by looking for the first item in the units array (ie. item with
index 0).

Fixed by includig the sequential flag as true in the API call.